fixed another warning

+++ main/lib/plugins/PALPIDEFSConverterPlugin.cc
@@ -1106,7 +1106,7 @@ namespace eudaq {
       for (unsigned int i_sector = 0; i_sector < n_sectors; ++i_sector) {
         if (successful_fits[sector] > 0) {
           (*thr_rms)[i_sector] = TMath::Sqrt(
-              (*thr_rms)[i_sector] / (double)successful_fits[i_sector] -
+              (*thr_rms)[i_sector] / (float)successful_fits[i_sector] -
               (*thr)[i_sector] * (*thr)[i_sector] /
                   (float)successful_fits[i_sector] /
                   (float)successful_fits[i_sector]);
